+++ src/main/java/com/example/itau/dtos/TransactionCreateDTO.java
@@ -1,7 +1,8 @@
 package com.example.itau.dtos;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
-public record TransactionCreateDTO(float valor, OffsetDateTime dataHora) {
+public record TransactionCreateDTO(BigDecimal valor, OffsetDateTime dataHora) {
 
 }
